added endpoints

+++ Backend/Database/Managers/UserManager.cs
@@ -6,7 +6,7 @@ namespace Backend.Database.Managers;
 public class UserManager
 {
     private readonly AppDbContext _context;
-    
+
     public UserManager(AppDbContext context)
     {
         _context = context;
@@ -26,21 +26,50 @@ public class UserManager
         return await _context.Users.FindAsync(id);
     }
 
-    public async Task Create(User user)
+    public async Task<bool> Create(User user)
     {
         _context.Users.Add(user);
-        await _context.SaveChangesAsync();
+        try
+        {
+            await _context.SaveChangesAsync();
+        }
+        catch (DbUpdateException)
+        {
+            return false;
+        }
+
+        return true;
     }
 
-    public async Task Modify(User user)
+    public async Task<bool> Modify(User user)
     {
         _context.Users.Update(user);
-        await _context.SaveChangesAsync();
+        try
+        {
+            await _context.SaveChangesAsync();
+        }
+        catch (DbUpdateException)
+        {
+            return false;
+        }
+
+        return true;
     }
-    
-    public async Task Delete(User user)
+
+    public async Task<bool> Delete(byte[] id)
     {
-        _context.Users.Remove(user);
-        await _context.SaveChangesAsync();
+        try
+        {
+            User? user = await _context.Users.FindAsync(id);
+            if (user == null) return true;
+            _context.Users.Remove(user);
+            await _context.SaveChangesAsync();
+        }
+        catch (DbUpdateException)
+        {
+            return false;
+        }
+
+        return true;
     }
 }
